moving batch_download.py and get_model.py to batch directory, and change get_model.py to _get_model.py
